pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulse-anthropic"
version = "0.1.0"
description = "Anthropic Claude adapter for PULSE Protocol — send PULSE messages, get Claude responses"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    {name = "PULSE Protocol Team", email = "pulse@protocol.org"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pulse-protocol>=0.5.0",
    "anthropic>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
]

[project.urls]
Homepage = "https://github.com/pulseprotocolorg-cyber/pulse-anthropic"
Repository = "https://github.com/pulseprotocolorg-cyber/pulse-anthropic"

[tool.setuptools.packages.find]
include = ["pulse_anthropic*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 100
